Fix and clean up package imports/deps
- Fix visibility of some functions/entities
- Restore autoexported vendor imports
- Clean up CryptoCore helpers
